dify-batch-processor/worker/requirements.txt
# Google Cloud
google-cloud-firestore

# HTTP Requests
requests
